Add in subcommand for compact.
Adds a subcommand to compact the log for debugging, and removed the appending of None LogEntry values in the event that they do occur during compaction.

This should not be an issue in practice as only live items exist in the index and the index is used for the compaction.